refactor(bootstrap): Move app to another file so it can be reused
